+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+#### 1.2.0
+* Fixed proper inclusion of maintenance mode when removed modifications.
+* Minor changes for compatibility.
+* Translated module name for the Russian locale.
+* Changed redirection when you save module settings.
+
 #### 1.1.0
 * Usability fixes
 * Success message moved to dropdown window